@@ -1,36 +1,61 @@
 ext {
 
-  compileSdkVersion = 27
-  buildToolsVersion = "27.1.0"
+  compileSdkVersion = 29
   minSdkVersion = 21
-  targetSdkVersion = 27
+  targetSdkVersion = 29
 
-  dagger_version = '2.4'
-  play_services_version = '12.0.1'
-  support_version = '27.1.0'
+  versions = [
+      'kotlin'             : '1.3.61',
+      'kotlinCoroutines'   : '1.3.3',
+      'kotlinSerialization': '0.14.0',
+      'ktor'               : '1.3.0',
+      'logback'            : '1.2.1',
+      'sqldelight'         : '1.2.2',
+      'dagger'             : '2.26',
+  ]
 
   libs = [
-      circularImageView : 'com.mikhaellopez:circularimageview:3.0.2',
-      constraintLayout  : "com.android.support.constraint:constraint-layout:1.0.2",
-      dagger            : "com.google.dagger:dagger:$dagger_version",
-      daggerCompiler    : "com.google.dagger:dagger-compiler:$dagger_version",
-      firebaseAuth      : "com.google.firebase:firebase-auth:$play_services_version",
-      firebaseCore      : "com.google.firebase:firebase-core:$play_services_version",
-      firebaseDatabase  : "com.google.firebase:firebase-database:$play_services_version",
-      firebaseStorage   : "com.google.firebase:firebase-storage:$play_services_version",
-      glide             : "com.github.bumptech.glide:glide:3.7.0",
-      junit4            : "junit:junit:4.12",
-      kotlinRuntime     : "org.jetbrains.kotlin:kotlin-runtime:$kotlin_version",
-      kotlinStdlib      : "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version",
-      kotlinTest        : "org.jetbrains.kotlin:kotlin-test:$kotlin_version",
-      mockitoCore       : "org.mockito:mockito-core:2.8.47",
-      mockitoKotlin     : "com.nhaarman:mockito-kotlin:1.5.0",
-      multilineToolbar  : "net.opacapp:multiline-collapsingtoolbar:27.1.0",
-      playServicesMaps  : "com.google.android.gms:play-services-maps:$play_services_version",
-      recyclerView     : "com.android.support:recyclerview-v7:$support_version",
-      supportAppCompat  : "com.android.support:appcompat-v7:$support_version",
-      supportCardView   : "com.android.support:cardview-v7:$support_version",
-      supportDesign     : "com.android.support:design:$support_version",
-      truth             : "com.google.truth:truth:0.39",
+      'dagger'         : [
+          'core'    : "com.google.dagger:dagger:${versions.dagger}",
+          'compiler': "com.google.dagger:dagger-compiler:${versions.dagger}",
+      ],
+      'androidx'       : [
+          'recyclerview'    : 'androidx.recyclerview:recyclerview:1.0.0',
+          'appcompat'       : 'androidx.appcompat:appcompat:1.0.0',
+          'cardview'        : 'androidx.cardview:cardview:1.0.0',
+          'constraintlayout': "androidx.constraintlayout:constraintlayout:2.0.0-beta4",
+          'fragment'        : "androidx.fragment:fragment-ktx:1.2.0",
+          'lifecycle'       : [
+              'runtime': "androidx.lifecycle:lifecycle-runtime-ktx:2.2.0"
+          ],
+          'viewpager'       : "androidx.viewpager2:viewpager2:1.0.0",
+      ],
+      'firebase'       : [
+          'auth'    : "com.google.firebase:firebase-auth:19.2.0",
+          'database': "com.google.firebase:firebase-database-ktx:19.2.1",
+          'storage' : "com.google.firebase:firebase-storage-ktx:19.1.1",
+      ],
+      'glide'            : "com.github.bumptech.glide:glide:3.7.0",
+      'insetter'       : "dev.chrisbanes:insetter-ktx:0.2.1",
+      'junit4'           : "junit:junit:4.12",
+      'kotlin'         : [
+          'stdlib'    : "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version",
+          'coroutines': [
+              'core'    : "org.jetbrains.kotlinx:kotlinx-coroutines-core:${versions.kotlinCoroutines}",
+              'android' : "org.jetbrains.kotlinx:kotlinx-coroutines-android:${versions.kotlinCoroutines}",
+              'firebase': "org.jetbrains.kotlinx:kotlinx-coroutines-play-services:${versions.kotlinCoroutines}",
+          ],
+          'test'      : "org.jetbrains.kotlin:kotlin-test:$kotlin_version",
+      ],
+      'mockitoCore'      : "org.mockito:mockito-core:2.8.47",
+      'mockitoKotlin'    : "com.nhaarman:mockito-kotlin:1.5.0",
+      'multilineToolbar' : "net.opacapp:multiline-collapsingtoolbar:27.1.1",
+      'playServicesMaps' : "com.google.android.gms:play-services-maps:17.0.0",
+      'material'       : 'com.google.android.material:material:1.2.0-alpha05',
+      'timber'         : [
+          'jdk'    : 'com.jakewharton.timber:timber-jdk:5.0.0-SNAPSHOT',
+          'android': 'com.jakewharton.timber:timber-android:5.0.0-SNAPSHOT',
+      ],
+      'truth'            : "com.google.truth:truth:0.39",
   ]
 }
