@@ -4,12 +4,12 @@ import android.content.Context
 import android.graphics.Canvas
 import android.graphics.Rect
 import android.graphics.drawable.Drawable
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 
 internal class DividerItemDecoration(private val context: Context) : RecyclerView.ItemDecoration() {
 
-    private val divider: Drawable
+    private val divider: Drawable?
 
     init {
         val a = context.obtainStyledAttributes(intArrayOf(android.R.attr.listDivider))
@@ -18,6 +18,8 @@ internal class DividerItemDecoration(private val context: Context) : RecyclerVie
     }
 
     override fun onDraw(c: Canvas, parent: RecyclerView, state: RecyclerView.State) {
+        if (divider == null) return
+
         val childCount = parent.childCount
         for (i in 0..childCount) {
             val view = parent.getChildAt(i)
@@ -36,6 +38,8 @@ internal class DividerItemDecoration(private val context: Context) : RecyclerVie
     }
 
     override fun getItemOffsets(outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State) {
+        if (divider == null) return
+
         val holder = parent.getChildViewHolder(view)
         if (holder is SessionAdapter.ViewHolder) {
             if (holder.adapterPosition != 0) {
