@@ -5,10 +5,10 @@ import android.content.res.Resources
 import android.graphics.Bitmap
 import android.graphics.Canvas
 import android.graphics.drawable.Drawable
-import android.support.annotation.CheckResult
-import android.support.annotation.DrawableRes
-import android.support.graphics.drawable.VectorDrawableCompat
-import android.support.v4.content.ContextCompat
+import androidx.annotation.CheckResult
+import androidx.annotation.DrawableRes
+import androidx.vectordrawable.graphics.drawable.VectorDrawableCompat
+import androidx.core.content.ContextCompat
 
 object DrawableUtils {
 
@@ -17,13 +17,11 @@ object DrawableUtils {
    * to a plain old drawable resource.
    */
   @CheckResult fun create(context: Context, @DrawableRes resourceId: Int): Drawable? {
-    var drawable: Drawable?
-    try {
-      drawable = VectorDrawableCompat.create(context.resources, resourceId, context.theme)
+    return try {
+      VectorDrawableCompat.create(context.resources, resourceId, context.theme)
     } catch (exception: Resources.NotFoundException) {
-      drawable = ContextCompat.getDrawable(context, resourceId)
+      ContextCompat.getDrawable(context, resourceId)
     }
-    return drawable
   }
 
 }
