@@ -1,67 +1,67 @@
 package com.chicagoroboto.data
 
-import android.app.Application
-import android.content.Context
-import android.content.SharedPreferences
+import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.database.DatabaseReference
 import com.google.firebase.database.FirebaseDatabase
 import com.google.firebase.storage.FirebaseStorage
 import com.google.firebase.storage.StorageReference
+import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import javax.inject.Singleton
 
 @Module
-class DataModule {
+abstract class DataModule {
 
-    @Provides @Singleton fun provideSharedPreferences(application: Application): SharedPreferences {
-        return application.getSharedPreferences("app_prefs", Context.MODE_PRIVATE)
+  @Binds
+  abstract fun bindFavoriteProvider(impl: FirebaseFavoriteProvider): FavoriteProvider
+
+  @Binds
+  abstract fun bindNotificationProvider(impl: LocalNotificationProvider): NotificationProvider
+
+  @Binds
+  @Singleton
+  abstract fun bindSessionDateProvider(impl: FirebaseSessionDateProvider): SessionDateProvider
+
+  @Binds
+  @Singleton
+  abstract fun bindSessionProvider(impl: FirebaseSessionProvider): SessionProvider
+
+  @Binds
+  @Singleton
+  abstract fun bindSpeakerProvider(impl: FirebaseSpeakerProvider): SpeakerProvider
+
+  @Binds
+  @Singleton
+  abstract fun bindVenueProvider(impl: FirebaseVenueProvider): VenueProvider
+
+  @Binds
+  @Singleton
+  abstract fun bindReviewProvider(impl: FirebaseFeedbackProvider): FeedbackProvider
+
+  @Binds
+  abstract fun bindAvatarProvider(impl: FirebaseAvatarProvider): AvatarProvider
+
+  @Binds
+  abstract fun bindUserProvider(impl: FirebaseUserProvider): UserProvider
+
+  companion object {
+
+    @Provides
+    @Singleton
+    fun provideDatabase(@EventId eventId: String): DatabaseReference {
+      FirebaseDatabase.getInstance().setPersistenceEnabled(true);
+      val eventRef = FirebaseDatabase.getInstance().reference.child("events").child(eventId)
+      eventRef.keepSynced(true)
+      return eventRef
     }
 
-    @Provides @Singleton fun provideFavoriteProvider(sharedPreferences: SharedPreferences): FavoriteProvider {
-        return LocalFavoriteProvider(sharedPreferences)
-    }
+    @Provides
+    @Singleton
+    fun provideStorage(@EventId eventId: String): StorageReference =
+        FirebaseStorage.getInstance().reference.child(eventId)
 
-    @Provides @Singleton fun providePreferencesProvider(application: Application): PreferencesProvider {
-        return PreferencesProvider(application)
-    }
-
-    @Provides @Singleton fun provideNotificationProvider(application: Application): NotificationProvider {
-        return LocalNotificationProvider(application)
-    }
-
-    @Provides @Singleton fun provideDatabase(@EventId eventId: String): DatabaseReference {
-        FirebaseDatabase.getInstance().setPersistenceEnabled(true);
-        val eventRef = FirebaseDatabase.getInstance().reference.child("events").child(eventId)
-        eventRef.keepSynced(true)
-        return eventRef
-    }
-
-    @Provides @Singleton fun provideStorage(@EventId eventId: String): StorageReference {
-        return FirebaseStorage.getInstance().reference.child(eventId)
-    }
-
-    @Provides @Singleton fun provideSessionDateProvider(db: DatabaseReference): SessionDateProvider {
-        return FirebaseSessionDateProvider(db)
-    }
-
-    @Provides @Singleton fun provideSessionProvider(db: DatabaseReference): SessionProvider {
-        return FirebaseSessionProvider(db)
-    }
-
-    @Provides @Singleton fun provideSpeakerProvider(db: DatabaseReference): SpeakerProvider {
-        return FirebaseSpeakerProvider(db)
-    }
-
-    @Provides @Singleton fun provideVenueProvider(db: DatabaseReference): VenueProvider {
-        return FirebaseVenueProvider(db)
-    }
-
-    @Provides @Singleton fun provideReviewProvider(db: DatabaseReference, preferencesProvider: PreferencesProvider): FeedbackProvider {
-        return FirebaseFeedbackProvider(db, preferencesProvider)
-    }
-
-    @Provides fun provideAvatarProvider(storageReference: StorageReference): AvatarProvider {
-        return FirebaseAvatarProvider(storageReference)
-    }
+    @Provides
+    fun provideAuth(): FirebaseAuth = FirebaseAuth.getInstance()
+  }
 }
