@@ -2,38 +2,29 @@ package com.chicagoroboto.data
 
 import com.chicagoroboto.model.Venue
 import com.google.firebase.database.*
+import kotlinx.coroutines.channels.awaitClose
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.channelFlow
+import timber.log.Timber
+import timber.log.error
+import javax.inject.Inject
 
-class FirebaseVenueProvider(val database: DatabaseReference) : VenueProvider {
+class FirebaseVenueProvider @Inject constructor(
+    val database: DatabaseReference
+) : VenueProvider {
 
-    private val queries: MutableMap<Any, Query> = mutableMapOf()
-    private val listeners: MutableMap<Any, ValueEventListener> = mutableMapOf()
+  override fun venue(): Flow<Venue> = channelFlow {
+    val query = database.child("venue")
+    val listener = query.addValueEventListener(object : ValueEventListener {
+      override fun onDataChange(data: DataSnapshot) {
+        offer(data.toVenue())
+      }
 
-    override fun addVenueListener(key: Any, onComplete: (Venue?) -> Unit) {
-        if (queries[key] != null) {
-            removeVenueListener(queries)
-        }
+      override fun onCancelled(error: DatabaseError) {
+        // no op
+      }
+    })
 
-        val listener = object : ValueEventListener {
-            override fun onDataChange(data: DataSnapshot?) {
-                onComplete(data?.getValue(Venue::class.java))
-            }
-
-            override fun onCancelled(e: DatabaseError?) {
-                onComplete(null)
-            }
-        }
-        listeners[key] = listener
-
-        val query = this.database.child("venue")
-        query.addValueEventListener(listener)
-        queries[key] = query
-    }
-
-    override fun removeVenueListener(key: Any) {
-        val query = queries[key]
-        query?.removeEventListener(listeners[key])
-
-        queries.remove(key)
-        listeners.remove(key)
-    }
+    awaitClose { query.removeEventListener(listener) }
+  }
 }
