@@ -1,13 +1,16 @@
 package com.chicagoroboto.features.main
 
+import com.chicagoroboto.data.UserProvider
 import com.chicagoroboto.features.location.LocationComponent
 import com.chicagoroboto.features.sessions.SessionListComponent
 import com.chicagoroboto.features.speakerlist.SpeakerListComponent
 import dagger.Subcomponent
 
-@Subcomponent(modules = arrayOf(MainModule::class))
+@Subcomponent(modules = [MainModule::class])
 interface MainComponent {
-    fun sessionListComponent(): SessionListComponent
-    fun speakerListComponent(): SpeakerListComponent
-    fun locationComponent(): LocationComponent
+  val userProvider: UserProvider
+  val speakerListComponentFactory: SpeakerListComponent.Factory
+
+  fun sessionListComponent(): SessionListComponent
+  fun locationComponent(): LocationComponent
 }
