@@ -1,16 +1,21 @@
 package com.chicagoroboto.data
 
+import kotlinx.coroutines.flow.Flow
+
 interface FavoriteProvider {
+
+  fun favorites(userId: String): Flow<Set<String>>
+
   /**
    * Adds a listener to be updated with the user's current list of favorite session ids.
    */
-  fun addFavoriteListener(key: String, onComplete: (sessions: Set<String>) -> Unit)
+  fun addFavoriteListener(userId: String, onComplete: (sessions: Set<String>) -> Unit)
 
   /**
    * Removes the listener that was added with <code>key</code>
    */
-  fun removeFavoriteListener(key: String)
+  fun removeFavoriteListener(userId: String)
 
-  fun addFavoriteSession(id: String)
-  fun removeFavoriteSession(id: String)
+  fun addFavoriteSession(userId: String, id: String)
+  fun removeFavoriteSession(userId: String, id: String)
 }
