@@ -1,7 +1,8 @@
 package com.chicagoroboto.data
 
 import com.chicagoroboto.model.Library
+import kotlinx.coroutines.flow.Flow
 
 interface LibraryProvider {
-    fun libraries(onComplete: (libraries: List<Library>) -> Unit)
+  fun libraries(): Flow<List<Library>>
 }
