@@ -1,6 +1,7 @@
 package com.chicagoroboto.data
 
+import kotlinx.coroutines.flow.Flow
+
 interface SessionDateProvider {
-    fun addSessionDateListener(key: Any, onComplete: (sessionDates: List<String>?) -> Unit)
-    fun removeSessionDateListener(key: Any)
+  fun sessionDates(): Flow<List<String>>
 }
