@@ -1,9 +1,10 @@
 package com.chicagoroboto.data
 
 import com.chicagoroboto.model.Speaker
+import kotlinx.coroutines.flow.Flow
 
 interface SpeakerProvider {
-    fun addSpeakerListener(key: Any, onComplete: (speakers: Map<String, Speaker>?) -> Unit)
-    fun addSpeakerListener(id: String, onComplete: (speaker: Speaker?) -> Unit)
-    fun removeSpeakerListener(key: Any)
+  fun speakers(): Flow<List<Speaker>>
+  fun speaker(speakerId: String): Flow<Speaker>
+  suspend fun avatar(speakerId: String): String
 }
