@@ -1,25 +1,63 @@
 package com.chicagoroboto.features.info
 
 import com.chicagoroboto.data.LibraryProvider
+import com.chicagoroboto.features.info.InfoPresenter.Event
+import com.chicagoroboto.features.info.InfoPresenter.Event.ClickedLibrary
+import com.chicagoroboto.features.info.InfoPresenter.Model
+import com.chicagoroboto.features.shared.Presenter
 import com.chicagoroboto.model.Library
 import com.chicagoroboto.navigator.WebNavigator
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.channels.ConflatedBroadcastChannel
+import kotlinx.coroutines.channels.SendChannel
+import kotlinx.coroutines.channels.consumeEach
+import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.asFlow
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.launch
+import javax.inject.Inject
 
-class InfoPresenter(val libraryProvider: LibraryProvider, val webNavigator: WebNavigator) : InfoMvp.Presenter {
+class InfoPresenter @Inject constructor(
+    private val libraryProvider: LibraryProvider,
+    private val webNavigator: WebNavigator
+) : Presenter<Model, Event> {
 
-    private var view: InfoMvp.View? = null
+  private val _models = ConflatedBroadcastChannel<Model>()
+  override val models: Flow<Model> get() = _models.asFlow()
 
-    override fun onAttach(view: InfoMvp.View) {
-        this.view = view
-        libraryProvider.libraries { libraries ->
-            this.view?.showLibraries(libraries)
+  private val _events = Channel<Event>(Channel.BUFFERED)
+  override val events: SendChannel<Event> get() = _events
+
+  override suspend fun start() = coroutineScope<Unit> {
+    var model = Model()
+    fun sendModel(newModel: Model) {
+      model = newModel
+      _models.offer(model)
+    }
+
+    launch {
+      libraryProvider.libraries().collect {
+        sendModel(model.copy(libraries = it))
+      }
+    }
+
+    launch {
+      _events.consumeEach { event ->
+        when (event) {
+          is ClickedLibrary -> {
+            webNavigator.navigateToUrl(event.library.url)
+          }
         }
+      }
     }
+  }
 
-    override fun onDetach() {
-        this.view = null
-    }
+  data class Model(
+      val libraries: List<Library> = emptyList()
+  )
 
-    override fun onLibraryClicked(library: Library) {
-        webNavigator.navigateToUrl(library.url)
-    }
+  sealed class Event {
+    data class ClickedLibrary(val library: Library) : Event()
+  }
 }
