@@ -1,11 +1,12 @@
 package com.chicagoroboto.model
 
-class Venue(val name: String? = null,
-            val address: String? = null,
-            val city: String? = null,
-            val state: String? = null,
-            val zip: String? = null,
-            val phone: String? = null,
-            val latitude: Double? = null,
-            val longitude: Double? = null) {
-}
+class Venue(
+    val name: String = "",
+    val address: String = "",
+    val city: String = "",
+    val state: String = "",
+    val zip: String = "",
+    val phone: String = "",
+    val latitude: Double = 41.8339042,
+    val longitude: Double = -88.012144
+)
